Refactor UserHomepage to be named Dashboard
